feat: split into list.c
