+++ doc/gcc/gcc.md
@@ -19,6 +19,8 @@ vimtutor是个很好用的教程，对初学者非常友好，要使用它可在
 	vimtutor
 	```
 
+网上有人做了个 Vim 的游戏：[Game](https://vim-adventures.com/)。
+
 ### 启动和退出
 
 默认启动 Vim 可用 `vim` 命令，启动后会显示 Vim 的首页，此时未打开任何文件。
@@ -31,8 +33,10 @@ vimtutor是个很好用的教程，对初学者非常友好，要使用它可在
 用 `vim -o file1 file2` 命令对 file1 和 file2 上下分割，用 `vim -O file1 file2` 命令对 file1 和 file2 左右分割。
 用 `vim -p file1 file2` 命令可将多个文件各自在标签页中打开，标签页间的跳转可用 `gt` 或者 `:n` 和 `:N` 命令。
 
+**Ctrl-S是锁屏操作， Vim 中请不要用这个命令来保存文件。**
+
 退出 Vim 请先回到命令模式(注意不要在插入模式下尝试退出)，然后输入 `:q` 命令即可退出。
-如果打开了文件且文件发生修改，则需使用 `:wq` 命令保存并退出，或者使用 `:q!` 强制退出(不保存文件)。
+如果打开了文件且文件发生修改，则需使用 `:wq` 或 `Shift+zz` 命令保存并退出，或者使用 `:q!` 强制退出(不保存文件)。
 如果打开了多个文件，可以逐个退出(即先退出一个文件，再退出另一个……)，也可以使用 `:wqa` 或 `:qa!` 命令一次性退出。
 
 
@@ -97,6 +101,9 @@ Vim 有以下几种模式：
 - gg=G 自动缩进整个文件
 - % 光标移动到匹配的括号处
 - `. 光标移动到上次编辑的位置
+- zz 让光标所在的行居屏幕中央
+- zt 让光标所在的行居屏幕最上一行
+- zb 让光标所杂的行居屏幕最下一行
 
 ### 命令模式下的编辑
 
@@ -225,7 +232,15 @@ vim 的各种插件(一般用 VimL 语言编写，来自官网或 GitHub) taglis
 
 gcc -lm -g example.c -o example.exe
 
--g -L -I -l -D
+-g 
+
+-L 
+
+-I 
+
+-l 
+
+-D
 
 ### 编译过程
 
@@ -265,6 +280,8 @@ O0 O1 O2 O3 Os
 复杂的编译命令可写在 Makefile 中
 
 ### 环境变量
+
+man ld.so
 
 LD_LIBRARY_PATH
