+++ src/gcc/Makefile
@@ -1,7 +1,7 @@
 all: run.exe
 
-run.exe: list.h example.c
-	gcc -g -std=c99 -lm example.c -o run.exe
+run.exe: list.h list.c example.c
+	gcc -g -std=c99 -lm list.c example.c -o run.exe
 
 .PHONY: clean
